xpkg: Be more specific about kinds when parsing a workspace
Previously, the workspace parser just looked at the kind of an object to
determine whether to parse it as a known type (XRD, Composition, etc.) or
consider it an example. It's now common for the examples directory to include a
`Function` for installing functions into a cluster for testing; these
`Function`s aren't from the `meta.pkg.crossplane.io` group that xpkg cares
about, but the `pkg.crossplane.io` group.

Compare the full GVKs to ensure we only capture the resources we want.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>