dependency: Separate cache operations from the add command
Previously, `up dependency add` called without an argument would update the
package cache, and called with `-c` would clean the cache. Neither of those
operations really makes sense in `add`, so move them to their own
commands (`update-cache` and `clean-cache`).

While we're here, update the wording in the help to be project-centric and add
`dep` as an alias so I can stop making typos in `dependency`.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>