project: Introduce `up project push` (#67)
* project: Tag on-disk images produced by `up project build`

Including the tag in the image lets us find the images by tag when we read them
back in to push them. This doesn't really matter right now, but will be
important when we add embedded functions and have multiple images that need to
be linked together by annotations. Also, once we add embedded functions we can
produce a single package file including the functions and the configuration
using go-containerregistry's `MultiWrite` method.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* project: Introduce `up project push`

Add a command to push a built image from a project to a registry. For now, build
and push are always separate steps, but we could add a `--push` option to build
or a `--build` option to push to streamline this for users.

The code for the new command is largely copied from the existing `up xpkg push`,
which we'll remove eventually.

Fixes #29

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* fixup! project: Introduce `up project push`

Factor out annotate

* fixup! project: Introduce `up project push`

Improve repo name check

---------

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>