project: Add support for embedded functions to `up project build`
As part of building a project, find functions in the `functions/` directory (or
at the path configured in the project file), build them into function xpkgs, and
add them as dependencies on the configuration. Write the configuration xpkg and
all the function xpkgs together into a single output file, which can be consumed
by `up project push`. Change the extension we're using for the output file to
`.uppkg` to make it clear that it's not an upstream-compatible xpkg (`crossplane
xpkg push` can't push it).

For now, we support two kinds of functions:

1. Functions with a Dockerfile are built using `docker build`. We use the Docker
   client library for this rather than shelling out, but it still requires a
   running Docker daemon.
2. Functions with a kcl.mod are built by adding their code to a special KCL base
   image that takes its input from a static location. Building these does not
   require a Docker daemon.

Fixes #28

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>