project: Make the build cache more effective
Previously, we initialized the build cache in the KCL builder and used it to
construct the KCL base image, which we then built on top of. It turns out that
because of some details of go-containerregistry when you build on top of a
cached image, the cache becomes ineffective.

Construct the cache in the build command so that it gets used when we pull the
layers, which happens while writing packages to disk. This makes a big
difference especially for KCL-based functions since their base layers are large.

Toward #78

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>