project: Apply annotations to images only once
Before we push an image we need to apply the relevant annotations to its layers,
since the layer annotations are not stored in the on-disk package. This requires
building a new image out of the newly-annotated layers, which can be slow.

Previously, we were annotating each image twice: once when we loaded it from the
package file and again when generating the index. We have to do it as part of
generating the index so that the index is correct (accounting for other callers
that don't pre-annotate). By returning the annotated images when we do that we
can avoid annotating them beforehand.

Fixes #78

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>