projects: Use an upstreamable KCL base image
Previously, the base image we used for KCL functions was built from a modified
version of function-kcl that took its code input from a hard-coded source
directory and had its package metadata removed. This worked, but if we moved
forward with it we would have to redo the changes every time the upstream
function-kcl was updated.

Update the build code to use a base image that's still not upstream function-kcl
but has changes that could potentially be merged into upstream
function-kcl. This version allows an environment variable to be set with a
default input, and uses that default input if input is not provided to the
function. The build code needs to do two things differently to use this image as
the base:

1. Remove the annotated layers that are specific to function-kcl, namely the
   package metadata layer and the examples. To be future-proof, we also remove
   the schema layer(s).
2. Add the necessary environment variable to the image's configuration.

While we're here, change `up project build` to write out the images produced by
`BuildIndex`. These are functionally identical to the input images to
`BuildIndex`, but have layer annotations applied.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>