ctx: Always use the spaces ingress
Previously, when connecting to a disconnected space (or a group in a
disconnected space), we used the hub context's apiserver address and auth for
the returned context. We would like to make some APIs (e.g., query API)
available only via the ingress, so we want to use the ingress address for all
contexts regardless of whether spaces are disconnected, connected, or cloud.

Fixes upbound/spaces#1560

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>