xpkg: Add a test for the Docker function builder
Add a basic integration test for the Docker function builder that builds a
simple image (`FROM scratch` so we don't have to pull anything) and verifies
that the image looks correct.

This test revealed a bug in the builder implementation. We were passing the
errgroup's context to go-containerregistry's daemon package, which tried to use
said context when we fetch the layers of the image after building it. This
fails, since that context is closed. Use the daemon package without a context to
resolve this.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>