projects: Set the repository automatically in `up project init`
When the user runs `up project init`, fill in the repository in the generated
upbound.yaml using the project name and the account name from their active up
profile. If they don't have an up profile (have never logged in) put in an
obvious placeholder that will cause the build to fail, so they must fill it in
before proceeding. This will avoid people getting all the way to the push step
before realizing the default is `example/project-template`, which they can't
push to.

Fixes #94

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>