composition: Remove network dependency from composition tests
Previously, the tests for `up composition generate` were fetching packages from
marketplace, which meant they broke if network access wasn't available or if the
environment where they ran had invalid marketplace credentials configured.

Implement a new package resolver for use in tests that can serve multiple
packages from a filesystem. Use this with go:embed to hold all the packages we
need for the composition tests.

While we're here, add a tool (in `hack/`) that can be used to pull packages for
use as testdata. This tool removes the non-package layers from the packages, so
that for function and provider images in particular we don't need to keep the
large controller layers in git.

Fixes #118

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>