projects: Factor out project building so it can be re-used
We're planning to implement `up project run`, which will build and push a
project, then install it on a dev control plane. Previously, all the build logic
was contained in the build command, so it couldn't easily be re-used in other
commands. Move it into the `internal/project` package.

As part of this move, keep all terminal printing in the command by having the
build code produce status updates on a channel. Add a new utility library,
`internal/async`, to help with this kind of thing.

Toward #121

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>