projects: Factor out project pushing so it can be re-used
We're planning to implement `up project run`, which will build and push a
project, then install it on a dev control plane. Previously, all the push logic
was contained in the push command, so it couldn't easily be re-used in other
commands. Move it into the `internal/project` package.

Toward #121

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>