projects: Use configured registry endpoint in `up project init`
To improve the testing experience in our dev and staging environments, use the
registry endpoint configured in the Upbound context instead of hardcoding the
production endpoint (`xpkg.upbound.io`). This means that if you have `UP_DOMAIN`
set to work with the dev or staging environments, a new project will use the
corresponding registry by default.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>