project: Introduce `up project move` (#136)
One point of friction with embedded functions is that the function references in
compositions need to refer to the embedded functions using their repository
names, which are generated during project build based on the repository
specified in the project metadata. The build doesn't update compositions, so
changing the repository in the project metadata (or using the `--repository`
flag on the build) can easily result in broken compositions.

This isn't likely to be a problem for users working on their own projects in the
normal dev workflow: they'll set the repository when they start and always push
to the same place. But it's very awkward for working with example repositories:
if a user checks out one of our examples, then builds and pushes it to their own
xpkg repository, it won't work correctly.

Introduce a helper command called `up project move` that updates the repository
in the project metadata and also updates all embedded function references in
compositions to match. We can include this command as a step in our tutorials to
make it easy for users to work with an example project.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>