kcl: Fix permissions in KCL images and let Docker pull them (#141)
* kcl: Clear history of the base image when building KCL functions

Docker won't pull an image if the informational `history` section of the config
doesn't match the actual layers. Containerd doesn't care, so the images actually
work fine in clusters, but it's nice to be able to pull them with Docker for
testing and debugging. Clearing the history makes that work.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* kcl: Make source files writable at function runtime

KCL requires that the `kcl.mod.lock` file be writable when a function is
executed. Crossplane runs function pods using an unprivileged user (with
hard-coded UID 2000), so the file needs to be owned by that user rather than
whatever user owned it in the source repo.

Override the UIDs and GIDs when adding files to the tarball for KCL
functions. To make this ergonomic, refactor `FSToTar` a bit to take functional
options.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

---------

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>