projects: Cache schemas for local XRDs during build
During `up project build` we generate KCL and Python schemas for all XRDs in the
project and include them in the package. To let users conveniently build
compositions on top of their own XRDs, also add those generated schemas to the
local schema cache. Do this before we build embedded functions to make sure the
schemas are up-to-date before any functions that might use them are built.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>