projects: Introduce `up project run` (#145)
* projects: Introduce `up project run`

The `up project run` command installs a project on a development control plane
for testing, creating the control plane first if needed. For now, since we don't
have a way to side-load packages into a dev control plane, it builds and pushes
the project to Marketplace with a generated `v0.0.0` tag.

Leaving tests as a TODO for the moment; we should probably build them on top of
a bunch of fakes (fake API to create control planes, fake builder, fake pusher),
which will take some time to build. I have manually tested this pretty
extensively and it will get more exercise at the next LUOP session.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* project: Ensure existing control planes are dev MCPs in run

Avoid users accidentally using `up project run` on a production MCP by checking
the class before using it.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* project: Factor out image label calculation

This is used in both build and run, so move it to a common package.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* actions: Bump go and golangci-lint versions

Updating up-sdk-go bumped our module's required go version to 1.23. The old
version of golangci-lint doesn't work with go newer than 1.22.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

* Fix issues noticed by the new golangci-lint version

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>

---------

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>