projects: Validate build cache contents to avoid broken builds
To optimize project builds, we create an on-disk cache of image layers using
go-containerregistry's cache package. The upstream cache package validates that
layers are the correct size before returning them, but doesn't validate that
their contents is correct. When it returns a correctly sized but corrupt
layer (i.e., a layer whoses contents does not match its digest), writing images
out to disk or pushing them to a registry will fail with confusing messages.

Implement a cache wrapper that validates layer contents before returning the
layer, and deletes invalid layers from the cache. Ideally this will go
upstream (as part of their filesystem cache implementation), but for now we can
do it locally to avoid the problem.

Fixes #152

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>