ctx: Start at the top level if the current context is unusable
If your current kubectl context is an Upbound context, but it's connected to a
different account or a different environment (dev vs. prod), `up ctx` will fail
with an error instead of presenting a menu.

After deriving the current state, check that the derived state is usable (i.e.,
returns items) and revert to the top level menu if it's not. This will give
users the org selector rather than an error if they login to an account that has
access to different orgs (or when switching between dev and prod).

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>