Remove null creationTimestamp from marshaled yaml consistently
There are a bunch of places where we marshal resources to yaml and want to omit
the null `creationTimestamp` from the metadata. Previously we were doing it in
some places but not others, and doing it a couple of different ways.

Introduce a new internal yaml package that wraps the commonly-used functions
from the Kubernetes yaml package, but strips null creationTimestamps when
marshaling. Use it wherever we want to marshal tidy YAML.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>