xrd: Don't output the status fields of generated XRDs
As a side-effect of our improved yaml cleanup in #231, we started outputting the
status fields of XRDs when we generate them. This isn't really harmful - the
status doesn't get applied when Crossplane installs the package - but it's ugly.

Use the new `yaml.RemoveField` option to remove the status when marshaling the
XRD.

Fixes #244

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>